Converted the messageLog bootstrapper to typescript.
